+++ scriptworker/constants.py
@@ -64,7 +64,10 @@ DEFAULT_CONFIG = frozendict({
     "verify_cot_signature": False,
     "cot_job_type": "unknown",  # e.g., signing
     "cot_product": "firefox",
+    "cot_version": 2,
+    "min_cot_version": 1,
     "max_chain_length": 20,
+    "max_jsone_iterations": 10,
 
     # Specify a default gpg home other than ~/.gnupg
     "gpg_home": None,
@@ -87,6 +90,9 @@ DEFAULT_CONFIG = frozendict({
     "pubkey_path": "...",
     "privkey_path": "...",
     "my_email": "scriptworker@example.com",
+
+    "project_configuration_url": "https://hg.mozilla.org/build/ci-configuration/raw-file/default/projects.yml",
+    "pushlog_url": "{repo}/json-pushes?changeset={revision}&tipsonly=1&version=2&full=1",
 
     "chain_of_trust_hash_algorithm": "sha256",
     "cot_schema_path": os.path.join(os.path.dirname(__file__), "data", "cot_v1_schema.json"),
@@ -193,13 +199,9 @@ DEFAULT_CONFIG = frozendict({
         "path_regexes": (
             "^(?P<path>/mozilla-(central|unified))(/|$)",
             "^(?P<path>/integration/(autoland|fx-team|mozilla-inbound))(/|$)",
-            "^(?P<path>/releases/mozilla-(beta|release|esr52))(/|$)",
-            # XXX remove /projects/maple and birch when taskcluster relpro
-            #     migration is tier1 and landed on mozilla-central
-            # XXX remove /projects/jamun when we no longer run staging releases
-            #     from it
-            # XXX remove /projects/oak when we no longer test updates against it
-            "^(?P<path>/projects/(birch|maple|jamun|oak))(/|$)",
+            "^(?P<path>/releases/mozilla-(beta|release|esr\d+))(/|$)",
+            "^(?P<path>/projects/([A-Za-z0-9-]+))(/|$)",
+            "^(?P<path>/(try))(/|$)",
         ),
     }), ),