+++ scriptworker/test/test_integration.py
@@ -246,6 +246,10 @@ def test_temp_creds(event_loop, context_function):
     "name": "mozilla-beta linux64 en-US repackage signing",
     "index": "gecko.v2.mozilla-beta.nightly.latest.firefox.linux64-nightly-repackage-signing",
     "task_type": "signing",
+}, {
+    "name": "mozilla-release linux64 en-US repackage signing",
+    "index": "gecko.v2.mozilla-release.nightly.latest.firefox.linux64-nightly-repackage-signing",
+    "task_type": "signing",
 }))
 @pytest.mark.asyncio
 async def test_verify_production_cot(branch_context):